chore: gate the server runtime behind configure
